requirements.txt
aiosql==9.0
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
cffi==1.15.1
click==8.1.7
cryptography==41.0.4
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg==3.1.11
pycparser==2.21
pydantic==2.4.1
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
